pb-rust: fixup tests
